requirements.txt
# requirements.txt
pandas==2.0.0
pyyaml==6.0
sqlalchemy==1.4.39
pytest==7.3.1
great_expectations==0.15.50
apache-airflow==2.6.1
python-dotenv==1.0.0
